backend/requirements.txt
fastapi
uvicorn
litellm
python-dotenv
pydantic
supabase
letta
langgraph
